pyproject.toml
[project]
name = "king-context"
version = "0.1.0"
description = "Local-first, token-efficient documentation server for LLM context injection"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.4.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "firecrawl-py>=1.0.0",
    "httpx>=0.27.0",
    "exa-py>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=0.25.0",
]

[project.scripts]
king-context = "king_context.server:main"
king-scrape = "king_context.scraper.cli:main"
king-research = "king_context.research.cli:main"
kctx = "context_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/king_context", "src/context_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
